feat: Add an in-game HUD editor allowing users to drag, scale, and reset module positions.
