+++ build.gradle
@@ -20,6 +20,7 @@ dependencies {
     mappings "net.fabricmc:yarn:${project.yarn_mappings}"
     // Fabric Loader
     modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
+    modImplementation "net.fabricmc.fabric-api:fabric-api:${project.fabric_version}"
 
     // Mod Menu (Optional)
     modImplementation "com.terraformersmc:modmenu:7.2.2"
