+++ gradle.properties
@@ -1,5 +1,5 @@
 org.gradle.jvmargs=-Xmx3G
-mod_version = 1.0.8-1
+mod_version = 1.0.8-2
 maven_group = com.eymistaken.simplecps
 archives_base_name = eymistakens-hud
