+++ src/main/java/com/eymistaken/simplecps/api/EymistakenHudPlugin.java
@@ -1,0 +1,7 @@
+package com.eymistaken.simplecps.api;
+
+import com.eymistaken.simplecps.HudModuleManager;
+
+public interface EymistakenHudPlugin {
+    void registerHudModules(HudModuleManager manager);
+}
