+++ src/main/resources/assets/simplecps/lang/en_us.json
@@ -1,0 +1,4 @@
+{
+    "key.simplecps.open_editor": "HUD Editor",
+    "key.category.simplecps.categories.eymistaken": "Eymistaken's HUD"
+}
